double -> int

+++ ZeefvanEratosthenes.Test/UnitTest1.cs
@@ -9,7 +9,7 @@ namespace ZeefvanEratosthenes.Tests
     public class UnitTest1
     {
         ZeefvanEratosthenes.Priemgetallen _zeef = new Priemgetallen();
-        List<double> bekendePriemgetallen = new List<double>()
+        List<int> bekendePriemgetallen = new List<int>()
         {2,3,5,7,11,13,17,19,23,29
 ,31,37,41,43,47,53,59,61,67,71
 ,73,79,83,89,97,101,103,107,109,113
