+++ ZeefvanEratosthenes/Program.cs
@@ -4,7 +4,7 @@
 //Maak een gesorteerde lijst van alle getallen van 2 tot een zelf te kiezen maximum.
 var maximum = 10000;
 var priemgetallen = new ZeefvanEratosthenes.Priemgetallen();
-List<double> output = priemgetallen.Zoek(maximum);
+List<int> output = priemgetallen.Zoek(maximum);
 
 Console.WriteLine(String.Join(", ", output));
 Console.WriteLine(output.Count);
